requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
arxiv==2.1.3
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
cassandra-driver==3.29.2
cassio==0.1.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.20
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
durationpy==0.9
emoji==2.14.0
eval_type_backport==0.2.0
faiss-cpu==1.9.0
fastapi==0.115.5
feedparser==6.0.11
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.12.0
grpcio==1.67.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.17
langchain-groq==0.2.1
langchain-huggingface==0.1.2
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langchainhub==0.1.21
langdetect==1.0.9
langgraph==0.2.46
langgraph-checkpoint==2.0.3
langgraph-sdk==0.1.35
langsmith==0.1.142
layoutparser==0.3.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.13.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnx==1.17.0
onnxruntime==1.20.0
openai==1.54.4
opencv-python==4.10.0.84
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-asgi==0.49b1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-util-http==0.49b1
orjson==3.10.11
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.20.0
pillow==11.0.0
portalocker==3.0.0
posthog==3.7.0
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.13
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.17
python-oxmsg==0.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.41.2
streamlit==1.40.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.11
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
transformers==4.46.2
typer==0.13.0
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.5
unstructured-client==0.27.0
unstructured-inference==0.8.1
urllib3==2.2.3
uvicorn==0.32.0
watchdog==6.0.0
watchfiles==0.24.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.17.1
zipp==3.21.0